requirements.txt
discord.py==1.6
colorhash==1.0.3
requests==2.25.1
